Works in Python 3+

+++ src/frama_numpy.py
@@ -31,7 +31,7 @@ from matplotlib import pyplot as plt
 Length = 10000
 x = np.linspace(0,10 * np.pi, Length)
 Price = 2 * np.sin(x)
-Price[Length/2:Length] += 3
+Price[int(Length/2):Length] += 3
 Noise = np.random.randn(Price.shape[0]) # white noise
 InputPrice = Price + Noise
 batch = 10
